% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseMiraResi.R
\name{analyseMiraResi}
\alias{analyseMiraResi}
\title{Provides separate residual analyses for each complete-data model generated
by with 'with.mids' function in 'mice'.}
\usage{
analyseMiraResi(mira)
}
\arguments{
\item{mira}{An object of class 'mira' generated by the 'with.mids' function
in 'mice'}
}
\value{
Residual plots (normality, residuals vs. fitted values, lag plot)
Shapiro-Wilk normality test, Levene Test for homogeneity of variance, 
Runs test for randomness for each of the 'm' repeated complete-data models
}
\description{
Provides separate residual analyses for each complete-data model generated
by with 'with.mids' function in 'mice'.
}
