% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lm.nls.R
\name{as.lm}
\alias{as.lm}
\title{as.lm.nls}
\usage{
as.lm(object, ...)
}
\description{
Generates a gradient of predicted values from an nls model
to match formatting for lm models. 
See: http://www.leg.ufpr.br/~walmes/cursoR/ciaeear/as.lm.R
}
\author{
Walmes Marques Zeviani
}
