% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescStats.R
\name{getDescStats}
\alias{getDescStats}
\title{Calculates descriptive statistics (mean, median, standard deviation, 
minumum, maximum, and n)}
\usage{
getDescStats(data, variable)
}
\arguments{
\item{data}{A data frame}

\item{variable}{A column with a numerical variable of interest}
}
\value{
A dataframe with descriptive statistics
}
\description{
Calculates descriptive statistics (mean, median, standard deviation, 
minumum, maximum, and n)
}
\examples{
getDescStats(iris, Sepal.Length)

}
