% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpStartVal.R
\name{getExpStartVal}
\alias{getExpStartVal}
\title{Get starting values for an exponential model}
\usage{
getExpStartVal(x, y)
}
\arguments{
\item{x}{A vector or dataframe column of an independent variable}

\item{y}{A vector or dataframe column of a dependent variable}
}
\value{
list
}
\description{
Obtains approximate starting values for exponential model 
'y~a*exp(b*x)' 
by linearizing the exponential model to 
'ln(y)~ln(a)+bx' 
and transforming parameter 'a'. Values generated with this function
may not be apropriate for other nonlinear models.
}
