% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanImp.R
\name{getMeanImp}
\alias{getMeanImp}
\title{This function takes a multiply imputed dataset ('mids' object), groups 
imputed values by the independent variable and calculates the mean across 
all 'm' imputed datasets for univariate time series data.}
\usage{
getMeanImp(mids, data, xVarName, yVarName)
}
\arguments{
\item{mids}{A 'mids' objects generated by 'mice'}

\item{data}{Original univariate time series with missing values}

\item{xVarName}{A character string with the name of the predictor variable
in the original dataset}

\item{yVarName}{A character string with the name of the response variable in 
'mids' object}
}
\value{
A data frame with a column for mean response variables grouped by the
corresponding predictor variable value for which they were imputed
}
\description{
This function takes a multiply imputed dataset ('mids' object), groups 
imputed values by the independent variable and calculates the mean across 
all 'm' imputed datasets for univariate time series data.
}
