% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMiceR2.R
\name{getMiceR2}
\alias{getMiceR2}
\title{Returns a R^2 value for nonlinear least squares (nls) 
 regression. This version is for use with 'mice' and does
 not return adjusted R^2.
 Adapted from the R^2 calculation from 'easynls' by
 Emmanuel Arnhold: https://rdrr.io/cran/easynls/}
\usage{
getMiceR2(nls)
}
\arguments{
\item{nls}{An object of class 'nls'}
}
\value{
R^2 value
}
\description{
Returns a R^2 value for nonlinear least squares (nls) 
 regression. This version is for use with 'mice' and does
 not return adjusted R^2.
 Adapted from the R^2 calculation from 'easynls' by
 Emmanuel Arnhold: https://rdrr.io/cran/easynls/
}
