% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNlsConfInt.R
\name{getNlsConfInt}
\alias{getNlsConfInt}
\title{Get NLS confidence intervals}
\usage{
getNlsConfInt(nls, data, xVarName)
}
\arguments{
\item{nls}{An 'nls' object}

\item{data}{A dataframe with data used to fit model}

\item{xVarName}{A character string with the name of the predictor 
variable in the dataframe 'data'}
}
\value{
A data frame with fitted values, upper and lower bounds, and 
a column with values for predictor variable
}
\description{
Calculates 95% confidence intervals for 'nls' objects using the 'as.lm.nls'
function by Walmes Zediani of the Universidade Federal do Paraná
(http://www.leg.ufpr.br/~walmes/cursoR/ciaeear/as.lm.nls.R). Returns a 
data frame with upper and lower bounds, fitted values, and values for
predictor variable to facilitate plotting.
}
