% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNlsParams.R
\name{getNlsParams}
\alias{getNlsParams}
\title{Get NLS parameter estimates

Extracts parameter estimates from an 'nls' model object}
\usage{
getNlsParams(nls)
}
\arguments{
\item{nls}{An 'nls' model object}
}
\value{
A list of parameter estimates
}
\description{
Get NLS parameter estimates

Extracts parameter estimates from an 'nls' model object
}
