% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNlsR2.R
\name{getNlsR2}
\alias{getNlsR2}
\title{R^2 values for NLS models}
\usage{
getNlsR2(nls)
}
\arguments{
\item{nls}{An object of class 'nls'}
}
\value{
R^2 value and adjusted R^2 value
}
\description{
Returns R^2 and adjusted R^2 for nonlinear least squares (nls) 
models.

Adapted from the cuasi-R^2 calculation from 'easynls' by
Emmanuel Arnhold: https://rdrr.io/cran/easynls/
}
