% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPooledParams.R
\name{getPooledParams}
\alias{getPooledParams}
\title{Get Pooled Parameters}
\usage{
getPooledParams(mipo)
}
\arguments{
\item{mipo}{A 'mipo' object generated by 'mice'}
}
\value{
list
}
\description{
Extracts pooled parameter estimates from a 'mipo' object.
}
