% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPooledPreds.R
\name{getPooledPreds}
\alias{getPooledPreds}
\title{Calculates pooled predicted values from muliptly imputed repeated analyses
from MICE.}
\usage{
getPooledPreds(mira)
}
\arguments{
\item{mira}{A 'mira' object generated by MICE}
}
\value{
A dataframe with pooled predicted values
}
\description{
Calculates pooled predicted values from muliptly imputed repeated analyses
from MICE.
}
