% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPooledR2.R
\name{getPooledR2}
\alias{getPooledR2}
\title{Utility to calculate R^2  values by pooling values from 
multiply imputed repeated analyses. 
Based on van Buuren (https://rdrr.io/cran/mice/src/R/pool.r.squared.R)}
\usage{
getPooledR2(mira)
}
\arguments{
\item{mira}{An object of class 'mira' generated by 'mice'}
}
\value{
R^2 value
}
\description{
Utility to calculate R^2  values by pooling values from 
multiply imputed repeated analyses. 
Based on van Buuren (https://rdrr.io/cran/mice/src/R/pool.r.squared.R)
}
