% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPooledSE.R
\name{getPooledSE}
\alias{getPooledSE}
\title{Pooled standard error from 'mice'}
\usage{
getPooledSE(mira)
}
\arguments{
\item{mira}{A 'mira' object generated by package 'mice'}
}
\value{
A numerical value (standard error)
}
\description{
Calculates and pools standard errors from multiply imputed 
repeated analyses, generated by package 'mice', according to Rubin's Rules. T
his function assumes normally distributed errors.
}
