% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRateComp.R
\name{getRateComp}
\alias{getRateComp}
\title{Calculates the predicted value at time 'n' using the model:}
\usage{
getRateComp(data, start_year, end_year, rate)
}
\arguments{
\item{data}{A data frame with one column for a "time" and a column for 
the dependent variable}

\item{start_year}{A numerical value corresponding to start year}

\item{end_year}{A numerical value corresponding to end year}

\item{rate}{A numerical value for the rate of change}
}
\value{
Prints predicted and observed value at time 'n'
}
\description{
The predicted value is then compared to the observed value for time 'n'
}
