% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsImpResiPlot.R
\name{obsImpResiPlot}
\alias{obsImpResiPlot}
\title{Plot residuals vs fitted for imputed and observed models}
\usage{
obsImpResiPlot(mira, nls)
}
\arguments{
\item{mira}{An object of class 'mira' generated by the 'with.mids' function 
in ' mice'}

\item{nls}{A fitted model object created with observed data only (e.g., an
nls model object created with observed dataset that has missing values).}
}
\value{
A plot of residual vs. fited values
}
\description{
Provides a plot with residuals and fitted values from all 
multiply imputed models (grey points) and from the observed data model 
(black points).

Based on method proposed by: 
Rizopoulos, D., Verbeke, G., & Molenberghs, G. (2010). Multiple-Imputation-
Based Residuals and Diagnostic Plots for Joint Models of Longitudinal and 
Survival Outcomes. Biometrics, 66(1), 20-29. 
https://doi.org/10.1111/j.1541-0420.2009.01273.x
}
