% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotImp.R
\name{pivotImp}
\alias{pivotImp}
\title{Pivot imputed data}
\usage{
pivotImp(mids, data, xVarName, yVarName)
}
\arguments{
\item{mids}{A 'mids' object generated by 'mice'}

\item{data}{Original dataset with missing values}

\item{xVarName}{A character string with the name of the predictor variable
in the original dataset}

\item{yVarName}{A character string with the name of the response variable in
the 'mids' object}
}
\value{
A data frame with three columns
}
\description{
This function takes a 'mids' object, generated by package 'mice', and 
creates a data frame with one long column for all imputed values, a column 
indicating the corresponding value of the predictor variable, and a column 
indicating to which of the 'm' imputed datasets it belongs. This function is 
useful for time-series data, where predictor variable (time) is 
complete and response variable values are missing.
}
