% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolResiduals.R
\name{poolResiduals}
\alias{poolResiduals}
\title{Analyse mean residuals from multiply imputed repeated analyses}
\usage{
poolResiduals(mira)
}
\arguments{
\item{mira}{An object of class 'mira' generated by the 'with.mids' function 
in ' mice'}
}
\value{
Residual plots (normality, residuals vs. fitted values, lag plot)
Shapiro-Wilk normality test, Levene Test for homogeneity of variance, 
Runs test for randomness.
}
\description{
Provides residual analysis for multiply imputed repeated analyses, generated
by package 'mice', by extracting residual and fitted values from each 
complete-data model generated by the 'with.mids' function in 'mice', and 
then averaging residual and fitted values across all models.

This functions averages residuals and fitted values averaged across all 'm' 
complete-data models, and provides a normality plot and a plot of residual 
vs. fitted values. It also provides a Shapiro-Wilk normality test, 
Levene's test for homogeneity of variance, and residual mean.
}
