% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cmydas.R
\name{theme_cmydas}
\alias{theme_cmydas}
\title{Custom ggplot theme for Conservation Letters}
\usage{
theme_cmydas(
  base_size = 22,
  base_family = "Lato",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_size}{Base fontsize}

\item{base_family}{Base font family}

\item{base_line_size}{Base line size}

\item{base_rect_size}{Base rect size}
}
\description{
Custom ggplot theme based on 'theme_light'
}
